Use new algo for dependencies
